@@ -3,31 +3,31 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  bloc: ^8.0.3
   cupertino_icons: ^1.0.4
   device_preview: ^1.0.0
   dio: ^4.0.6
   dartz: ^0.10.1
-  flutter_bloc: ^8.0.1
+  flutter_mobx: ^2.0.6+1
   flutter_modular: ^5.0.2
   flutter_svg: ^1.0.3
-  google_fonts: ^2.3.2
+  google_fonts: ^3.0.1
   json_annotation: ^4.5.0
   lottie: ^1.2.2
+  mobx: ^2.0.7+2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  bloc_test: ^9.0.3
   build_runner: ^2.1.10
   faker_dart: ^0.1.4
-  flutter_lints: ^1.0.4
+  flutter_lints: ^2.0.1
   json_serializable: ^6.2.0
+  mobx_codegen: ^2.0.7
 
   # For information on the generic Dart part of this file, see the
   # following page: https://dart.dev/tools/pub/pubspec
