@@ -1,25 +1,39 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
 import '../../../../../core/design_system/components/registration_group.component.dart';
 import '../../../../../core/design_system/widgets/input_text.widget.dart';
+import '../../stores/registration.store.dart';
 
-class RegistrationNamePage extends StatelessWidget {
-  RegistrationNamePage({Key? key}) : super(key: key);
-
-  final TextEditingController _controller = TextEditingController();
+class RegistrationNamePage extends StatefulWidget {
+  const RegistrationNamePage({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) => RegistrationGroupComponent(
+  State<RegistrationNamePage> createState() => _RegistrationNamePageState();
+}
+
+class _RegistrationNamePageState
+    extends ModularState<RegistrationNamePage, RegistrationStore> {
+  @override
+  Widget build(BuildContext context) {
+    return Observer(
+      builder: (_) => RegistrationGroupComponent(
         widget: InputTextWidget(
-          controller: _controller,
           hint: 'Nome do grupo',
+          onChange: (String text) {
+            store.setName(text);
+          },
         ),
         textTitle: 'Dê um nome ao seu grupo',
         textSubtitle: 'Esse nome será utilizado para se referir ao seu grupo.',
         titleButton: 'Continuar',
         titleTextButton: 'Voltar',
-        actionButton: () => Modular.to.pushNamed('/registration_modality'),
+        actionButton: store.name.length < 3
+            ? null
+            : () => Modular.to.pushNamed('/group/registration_modality'),
         actionTextButton: () => Modular.navigatorDelegate?.pop(),
-      );
+      ),
+    );
+  }
 }
